+++ web/src/main/java/com/homeapp/javatraining/exception/user/UserBlockedException.java
@@ -1,0 +1,10 @@
+package com.homeapp.javatraining.exception.user;
+
+import com.homeapp.javatraining.exception.ApiException;
+
+public class UserBlockedException extends ApiException {
+
+    public UserBlockedException() {
+        super("User is blocked");
+    }
+}